+++ crates/bevy_xr_utils/Cargo.toml
@@ -8,3 +8,5 @@ edition = "2021"
 [dependencies]
 bevy.workspace = true
 bevy_xr.path = "../bevy_xr"
+openxr = "0.18.0"
+bevy_openxr.path = "../bevy_openxr"
