+++ crates/bevy_xr_utils/src/lib.rs
@@ -1,3 +1,3 @@
 // use bevy::prelude::*;
 pub mod hand_gizmos;
-
+pub mod xr_utils_actions;
